fix(lib/util & lab/04): move resize stuff into lib/util
