+++ src/lib/util/resize.test.ts
@@ -1,0 +1,5 @@
+import resize from '@fartts/lib/util/resize';
+
+test('resize', () => {
+  expect(resize).toBeDefined();
+});
